GOTerms class moved to AnnotationDbi, GONode removed
In order for GO and GO.db packages to be used interchangeably and for
function in the annotate package to function properly, we must have a
single shared class representing GO terms.  This patch moves the
GOTerms class definition to AnnotationDbi and removes references to
the GONode class which was intended as a replacement.


git-svn-id: file:///home/git/hedgehog.fhcrc.org/bioconductor/trunk/madman/Rpacks/annotate@26854 bc3139a8-67e5-0310-9ffc-ced21a209358