Modifications to make PWAmat() work with new pkgs
Changes were to PWAmat() and enable it to get an incidence matrix for
the newer organism based annotations as well as the older chip based
packages.  Changes primarily are to allow PWAmat() to know when it has
an organism based package or chip based package so it can return the
correct result.



git-svn-id: file:///home/git/hedgehog.fhcrc.org/bioconductor/trunk/madman/Rpacks/annotate@29500 bc3139a8-67e5-0310-9ffc-ced21a209358