Fix accessionToUID() error on Windows
RCurl::getURL() started to fail recently on Windows for some HTTPS
requests e.g. we started to see the following error on all the
Bioconductor Windows build machines (tokay1, riesling1, tokay2):

  > library(RCurl)
  > query <- "https://www.ncbi.nlm.nih.gov/entrez/eutils/esearch.fcgi?db=gene&tool=bioconductor&term=U03397"
  > doc <- RCurl::getURL(query)
  Error in function (type, msg, asError = TRUE)  :
    error:1407742E:SSL routines:SSL23_GET_SERVER_HELLO:tlsv1 alert
    protocol version

The reason is not clear but we suspect a server-side issue with the
SSL certificate that only manifests itself with RCurl::getURL() on
Windows. httr::GET() doesn't seem to be affected so accessionToUID()
now uses this instead of RCurl::getURL().